Remove global mutable variables from hle core.
All global mutable variables of the HLE core were regrouped into struct hle_t.
This allow third party to reuse the HLE core, without limiting them to a single process.
All they have to do is to declare another struct hle_t and pass it to hle_execute.
For mupen64plus plugin, a global hle_t is declared and passed to hle_execute.

I would have expected a small performance hit, because of the added level of indirection,
but it doesn't seem to be really noticeable for me.