Implement RSP fallback.
This allows to specify an arbitrary RSP plugin to use when an unknown
ucode is encountered. It is particularly usefull when combined with an
LLE RSP plugin.

Sending audio lists or display lists to audio (resp. gfx) plugins can
also be specified through config parameters. A blacklist of known
problematic gfx ucodes is integrated (same as rsp-cxd4) to avoid sending
them to the gfx plugin.